ENT-5134 Introduce CordaSessionFactoryFactory interface and the H2 implememntation
