+++ node-api/src/main/kotlin/net/corda/nodeapi/internal/persistence/HibernateConfiguration.kt
@@ -37,26 +37,26 @@ class HibernateConfiguration(
     companion object {
         private val logger = contextLogger()
 
-        // register custom converters
-        fun buildHibernateMetadata(metadataBuilder: MetadataBuilder, jdbcUrl:String, attributeConverters: Collection<AttributeConverter<*, *>>): Metadata {
-            metadataBuilder.run {
-                attributeConverters.forEach { applyAttributeConverter(it) }
-                // Register a tweaked version of `org.hibernate.type.MaterializedBlobType` that truncates logged messages.
-                // to avoid OOM when large blobs might get logged.
-                applyBasicType(CordaMaterializedBlobType, CordaMaterializedBlobType.name)
-                applyBasicType(CordaWrapperBinaryType, CordaWrapperBinaryType.name)
-
-                // Create a custom type that will map a blob to byteA in postgres and as a normal blob for all other dbms.
-                // This is required for the Checkpoints as a workaround for the issue that postgres has on azure.
-                if (jdbcUrl.contains(":postgresql:", ignoreCase = true)) {
-                    applyBasicType(MapBlobToPostgresByteA, MapBlobToPostgresByteA.name)
-                } else {
-                    applyBasicType(MapBlobToNormalBlob, MapBlobToNormalBlob.name)
-                }
-
-                return build()
-            }
-        }
+//        // register custom converters
+//        fun buildHibernateMetadata(metadataBuilder: MetadataBuilder, jdbcUrl:String, attributeConverters: Collection<AttributeConverter<*, *>>): Metadata {
+//            metadataBuilder.run {
+//                attributeConverters.forEach { applyAttributeConverter(it) }
+//                // Register a tweaked version of `org.hibernate.type.MaterializedBlobType` that truncates logged messages.
+//                // to avoid OOM when large blobs might get logged.
+//                applyBasicType(CordaMaterializedBlobType, CordaMaterializedBlobType.name)
+//                applyBasicType(CordaWrapperBinaryType, CordaWrapperBinaryType.name)
+//
+//                // Create a custom type that will map a blob to byteA in postgres and as a normal blob for all other dbms.
+//                // This is required for the Checkpoints as a workaround for the issue that postgres has on azure.
+//                if (jdbcUrl.contains(":postgresql:", ignoreCase = true)) {
+//                    applyBasicType(MapBlobToPostgresByteA, MapBlobToPostgresByteA.name)
+//                } else {
+//                    applyBasicType(MapBlobToNormalBlob, MapBlobToNormalBlob.name)
+//                }
+//
+//                return build()
+//            }
+//        }
     }
 
     private val sessionFactories = cacheFactory.buildNamed<Set<MappedSchema>, SessionFactory>(Caffeine.newBuilder(), "HibernateConfiguration_sessionFactories")
@@ -70,35 +70,8 @@ class HibernateConfiguration(
     fun sessionFactoryForSchemas(key: Set<MappedSchema>): SessionFactory = sessionFactories.get(key, ::makeSessionFactoryForSchemas)!!
 
     private fun makeSessionFactoryForSchemas(schemas: Set<MappedSchema>): SessionFactory {
-        logger.info("Creating session factory for schemas: $schemas")
-        val serviceRegistry = BootstrapServiceRegistryBuilder().build()
-        val metadataSources = MetadataSources(serviceRegistry)
-
-        val hbm2dll: String =
-        if(databaseConfig.initialiseSchema && databaseConfig.initialiseAppSchema == SchemaInitializationType.UPDATE) {
-            "update"
-        } else if((!databaseConfig.initialiseSchema && databaseConfig.initialiseAppSchema == SchemaInitializationType.UPDATE)
-                || databaseConfig.initialiseAppSchema == SchemaInitializationType.VALIDATE) {
-            "validate"
-        } else {
-            "none"
-        }
-
-        // We set a connection provider as the auto schema generation requires it.  The auto schema generation will not
-        // necessarily remain and would likely be replaced by something like Liquibase.  For now it is very convenient though.
-        val config = Configuration(metadataSources).setProperty("hibernate.connection.provider_class", NodeDatabaseConnectionProvider::class.java.name)
-                .setProperty("hibernate.format_sql", "true")
-                .setProperty("hibernate.hbm2ddl.auto", hbm2dll)
-                .setProperty("javax.persistence.validation.mode", "none")
-                .setProperty("hibernate.connection.isolation", databaseConfig.transactionIsolationLevel.jdbcValue.toString())
-
-        schemas.forEach { schema ->
-            // TODO: require mechanism to set schemaOptions (databaseSchema, tablePrefix) which are not global to session
-            schema.mappedTypes.forEach { config.addAnnotatedClass(it) }
-        }
-
-        val sessionFactory = buildSessionFactory(config, metadataSources, customClassLoader)
-        logger.info("Created session factory for schemas: $schemas")
+        val sessionFactoryFactory = H2SessionFactoryFactory()
+        val sessionFactory = sessionFactoryFactory.makeSessionFactoryForSchemas(databaseConfig, schemas, customClassLoader, attributeConverters)
 
         // export Hibernate JMX statistics
         if (databaseConfig.exportHibernateJMXStatistics)
@@ -120,27 +93,6 @@ class HibernateConfiguration(
             mbeanServer.registerMBean(statisticsMBean, statsName)
         } catch (e: Exception) {
             logger.warn(e.message)
-        }
-    }
-
-    private fun buildSessionFactory(config: Configuration, metadataSources: MetadataSources, customClassLoader: ClassLoader?): SessionFactory {
-        config.standardServiceRegistryBuilder.applySettings(config.properties)
-
-        if (customClassLoader != null) {
-            config.standardServiceRegistryBuilder.addService(
-                    ClassLoaderService::class.java,
-                    ClassLoaderServiceImpl(customClassLoader))
-        }
-
-        @Suppress("DEPRECATION")
-        val metadataBuilder = metadataSources.getMetadataBuilder(config.standardServiceRegistryBuilder.build())
-        val metadata = buildHibernateMetadata(metadataBuilder, jdbcUrl, attributeConverters)
-        return metadata.sessionFactoryBuilder.run {
-            allowOutOfTransactionUpdateOperations(true)
-            applySecondLevelCacheSupport(false)
-            applyQueryCacheSupport(false)
-            enableReleaseResourcesOnCloseEnabled(true)
-            build()
         }
     }
 
@@ -168,41 +120,6 @@ class HibernateConfiguration(
         override fun isUnwrappableAs(unwrapType: Class<*>?): Boolean = unwrapType == NodeDatabaseConnectionProvider::class.java
     }
 
-    // A tweaked version of `org.hibernate.type.MaterializedBlobType` that truncates logged messages.  Also logs in hex.
-    object CordaMaterializedBlobType : AbstractSingleColumnStandardBasicType<ByteArray>(BlobTypeDescriptor.DEFAULT, CordaPrimitiveByteArrayTypeDescriptor) {
-        override fun getName(): String {
-            return "materialized_blob"
-        }
-    }
-
-    // A tweaked version of `org.hibernate.type.descriptor.java.PrimitiveByteArrayTypeDescriptor` that truncates logged messages.
-    private object CordaPrimitiveByteArrayTypeDescriptor : PrimitiveByteArrayTypeDescriptor() {
-        private const val LOG_SIZE_LIMIT = 1024
-
-        override fun extractLoggableRepresentation(value: ByteArray?): String {
-            return if (value == null) {
-                super.extractLoggableRepresentation(value)
-            } else {
-                if (value.size <= LOG_SIZE_LIMIT) {
-                    "[size=${value.size}, value=${value.toHexString()}]"
-                } else {
-                    "[size=${value.size}, value=${value.copyOfRange(0, LOG_SIZE_LIMIT).toHexString()}...truncated...]"
-                }
-            }
-        }
-    }
-
-    // A tweaked version of `org.hibernate.type.WrapperBinaryType` that deals with ByteArray (java primitive byte[] type).
-    object CordaWrapperBinaryType : AbstractSingleColumnStandardBasicType<ByteArray>(VarbinaryTypeDescriptor.INSTANCE, PrimitiveByteArrayTypeDescriptor.INSTANCE) {
-        override fun getRegistrationKeys(): Array<String> {
-            return arrayOf(name, "ByteArray", ByteArray::class.java.name)
-        }
-
-        override fun getName(): String {
-            return "corda-wrapper-binary"
-        }
-    }
-
     // Maps to a byte array on postgres.
     object MapBlobToPostgresByteA : AbstractSingleColumnStandardBasicType<ByteArray>(VarbinaryTypeDescriptor.INSTANCE, PrimitiveByteArrayTypeDescriptor.INSTANCE) {
         override fun getRegistrationKeys(): Array<String> {
@@ -214,9 +131,4 @@ class HibernateConfiguration(
         }
     }
 
-    object MapBlobToNormalBlob : MaterializedBlobType() {
-        override fun getName(): String {
-            return "corda-blob"
-        }
-    }
 }
